cmd/snap-update-ns: move C parts to separate file
